+++ apps/router-e2e/app.config.js
@@ -41,8 +41,7 @@ module.exports = {
             : process.env.E2E_ROUTER_ASYNC === 'false'
             ? false
             : process.env.E2E_ROUTER_ASYNC || false,
-        // THIS WILL BREAK YOUR APP -- DO NOT USE!
-        unstable_src: path.join('__e2e__', process.env.E2E_ROUTER_SRC ?? 'static-rendering', 'app'),
+        root: path.join('__e2e__', process.env.E2E_ROUTER_SRC ?? 'static-rendering', 'app'),
         origin: 'https://smart-symbiote.netlify.app/',
       },
     ],