+++ apps/router-e2e/metro.config.js
@@ -1,5 +1,4 @@
 const { createMetroConfiguration } = require('expo-yarn-workspaces');
-const { FileStore } = require('@expo/metro-config/file-store');
 
 // Find the project and workspace directories
 const projectRoot = __dirname;
@@ -19,15 +18,6 @@ config.resolver.blockList = [
   /\/expo-router\/node_modules\/@react-navigation/,
   /node_modules\/@react-navigation\/native-stack\/node_modules\/@react-navigation\//,
   /node_modules\/pretty-format\/node_modules\/react-is/,
-];
-
-config.cacheStores = [
-  // Ensure the cache isn't shared between projects
-  // this ensures the transform-time environment variables are changed to reflect
-  // the current project.
-  new FileStore({
-    root: path.join(projectRoot, 'node_modules/.cache/metro', process.env.E2E_ROUTER_SRC || 'app'),
-  }),
 ];
 
 module.exports = config;