+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -17,7 +17,7 @@ import { createRouteHandlerMiddleware } from './createServerRouteMiddleware';
 import { ExpoRouterServerManifestV1, fetchManifest } from './fetchRouterManifest';
 import { instantiateMetroAsync } from './instantiateMetro';
 import { metroWatchTypeScriptFiles } from './metroWatchTypeScriptFiles';
-import { getRouterDirectoryWithManifest, isApiRouteConvention } from './router';
+import { getRouterDirectoryModuleIdWithManifest, isApiRouteConvention } from './router';
 import { serializeHtmlWithAssets } from './serializeHtml';
 import { observeApiRouteChanges, observeFileChanges } from './waitForMetroToObserveTypeScriptFile';
 import { ExportAssetMap } from '../../../export/saveAssets';
@@ -95,18 +95,19 @@ export class MetroBundlerDevServer extends BundlerDevServer {
 
   async exportExpoRouterApiRoutesAsync({
     mode,
-    appDir,
     outputDir,
     prerenderManifest,
     baseUrl,
+    routerRoot,
   }: {
     mode: 'development' | 'production';
-    appDir: string;
     outputDir: string;
     // This does not contain the API routes info.
     prerenderManifest: ExpoRouterServerManifestV1;
     baseUrl: string;
+    routerRoot: string;
   }): Promise<{ files: ExportAssetMap; manifest: ExpoRouterServerManifestV1<string> }> {
+    const appDir = path.join(this.projectRoot, routerRoot);
     const manifest = await this.getExpoRouterRoutesManifestAsync({ appDir });
 
     const files: ExportAssetMap = new Map();
@@ -115,7 +116,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       const filepath = path.join(appDir, route.file);
       const contents = await bundleApiRoute(this.projectRoot, filepath, {
         mode,
-        appDir,
+        routerRoot,
         port: this.getInstance()?.location.port,
         shouldThrow: true,
         baseUrl,
@@ -161,10 +162,12 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     mode,
     minify = mode !== 'development',
     baseUrl,
+    routerRoot,
   }: {
     mode: 'development' | 'production';
     minify?: boolean;
     baseUrl: string;
+    routerRoot: string;
   }): Promise<{
     serverManifest: ExpoRouterServerManifestV1;
     manifest: ExpoRouterRuntimeManifest;
@@ -179,6 +182,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
         // Ensure the API Routes are included
         environment: 'node',
         baseUrl,
+        routerRoot,
       });
 
     return {
@@ -200,6 +204,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     mainModuleName,
     isExporting,
     asyncRoutes,
+    routerRoot,
   }: {
     isExporting: boolean;
     mode: string;
@@ -208,6 +213,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     baseUrl?: string;
     mainModuleName?: string;
     asyncRoutes: boolean;
+    routerRoot: string;
   }): Promise<{ artifacts: SerialAsset[]; assets?: AssetData[] }> {
     const devBundleUrlPathname = createBundleUrlPath({
       platform: 'web',
@@ -222,6 +228,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       asyncRoutes,
       baseUrl,
       isExporting,
+      routerRoot,
     });
 
     const bundleUrl = new URL(devBundleUrlPathname, this.getDevServerUrl()!);
@@ -289,6 +296,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       mode,
       minify = mode !== 'development',
       baseUrl,
+      routerRoot,
       isExporting,
       asyncRoutes,
     }: {
@@ -297,6 +305,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       minify?: boolean;
       baseUrl: string;
       asyncRoutes: boolean;
+      routerRoot: string;
     }
   ) {
     const devBundleUrlPathname = createBundleUrlPath({
@@ -308,6 +317,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       baseUrl,
       isExporting,
       asyncRoutes,
+      routerRoot,
     });
 
     const bundleStaticHtml = async (): Promise<string> => {
@@ -320,6 +330,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
           // Ensure the API Routes are included
           environment: 'node',
           baseUrl,
+          routerRoot,
         }
       );
 
@@ -328,7 +339,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     };
 
     const [{ artifacts: resources }, staticHtml] = await Promise.all([
-      this.getStaticResourcesAsync({ isExporting, mode, minify, baseUrl, asyncRoutes }),
+      this.getStaticResourcesAsync({ isExporting, mode, minify, baseUrl, asyncRoutes, routerRoot }),
       bundleStaticHtml(),
     ]);
     const content = serializeHtmlWithAssets({
@@ -456,12 +467,14 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       if (useServerRendering) {
         const baseUrl = getBaseUrlFromExpoConfig(exp);
         const asyncRoutes = getAsyncRoutesFromExpoConfig(exp, options.mode ?? 'development', 'web');
-        const appDir = getRouterDirectoryWithManifest(this.projectRoot, exp);
+        const routerRoot = getRouterDirectoryModuleIdWithManifest(this.projectRoot, exp);
+        const appDir = path.join(this.projectRoot, routerRoot);
         middleware.use(
           createRouteHandlerMiddleware(this.projectRoot, {
             ...options,
             appDir,
             baseUrl,
+            routerRoot,
             getWebBundleUrl: manifestMiddleware.getWebBundleUrl.bind(manifestMiddleware),
             getStaticPageAsync: (pathname) => {
               return this.getStaticPageAsync(pathname, {
@@ -470,6 +483,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
                 minify: options.minify,
                 baseUrl,
                 asyncRoutes,
+                routerRoot,
               });
             },
           })
@@ -490,7 +504,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
                 if (op === 'change' || op === 'add') {
                   rebundleApiRoute(this.projectRoot, filepath, {
                     ...options,
-                    appDir,
+                    routerRoot,
                     baseUrl,
                   });
                 }