+++ packages/@expo/cli/src/start/server/middleware/metroOptions.ts
@@ -3,20 +3,9 @@ import type { BundleOptions as MetroBundleOptions } from 'metro/src/shared/types
 import resolveFrom from 'resolve-from';
 
 import { env } from '../../../utils/env';
+import { getRouterDirectoryModuleIdWithManifest } from '../metro/router';
 
 const debug = require('debug')('expo:metro:options') as typeof console.log;
-
-export function shouldEnableAsyncImports(projectRoot: string): boolean {
-  if (env.EXPO_NO_METRO_LAZY) {
-    return false;
-  }
-
-  // `@expo/metro-runtime` includes support for the fetch + eval runtime code required
-  // to support async imports. If it's not installed, we can't support async imports.
-  // If it is installed, the user MUST import it somewhere in their project.
-  // Expo Router automatically pulls this in, so we can check for it.
-  return resolveFrom.silent(projectRoot, '@expo/metro-runtime') != null;
-}
 
 export type ExpoMetroOptions = {
   platform: string;
@@ -34,7 +23,31 @@ export type ExpoMetroOptions = {
 
   baseUrl?: string;
   isExporting: boolean;
+  /** Module ID relative to the projectRoot for the Expo Router app directory. */
+  routerRoot: string;
 };
+
+export type SerializerOptions = {
+  includeSourceMaps?: boolean;
+  includeBytecode?: boolean;
+  output?: 'static';
+};
+
+export type ExpoMetroBundleOptions = MetroBundleOptions & {
+  serializerOptions?: SerializerOptions;
+};
+
+export function shouldEnableAsyncImports(projectRoot: string): boolean {
+  if (env.EXPO_NO_METRO_LAZY) {
+    return false;
+  }
+
+  // `@expo/metro-runtime` includes support for the fetch + eval runtime code required
+  // to support async imports. If it's not installed, we can't support async imports.
+  // If it is installed, the user MUST import it somewhere in their project.
+  // Expo Router automatically pulls this in, so we can check for it.
+  return resolveFrom.silent(projectRoot, '@expo/metro-runtime') != null;
+}
 
 function withDefaults({
   mode = 'development',
@@ -51,16 +64,6 @@ function withDefaults({
     ...props,
   };
 }
-
-export type SerializerOptions = {
-  includeSourceMaps?: boolean;
-  includeBytecode?: boolean;
-  output?: 'static';
-};
-
-export type ExpoMetroBundleOptions = MetroBundleOptions & {
-  serializerOptions?: SerializerOptions;
-};
 
 export function getBaseUrlFromExpoConfig(exp: ExpoConfig) {
   return exp.experiments?.baseUrl?.trim().replace(/\/+$/, '') ?? '';
@@ -80,6 +83,18 @@ export function getAsyncRoutesFromExpoConfig(exp: ExpoConfig, mode: string, plat
   return [mode, true].includes(asyncRoutesSetting);
 }
 
+export function getMetroDirectBundleOptionsForExpoConfig(
+  projectRoot: string,
+  exp: ExpoConfig,
+  options: Omit<ExpoMetroOptions, 'baseUrl' | 'routerRoot'>
+): Partial<ExpoMetroBundleOptions> {
+  return getMetroDirectBundleOptions({
+    ...options,
+    baseUrl: getBaseUrlFromExpoConfig(exp),
+    routerRoot: getRouterDirectoryModuleIdWithManifest(projectRoot, exp),
+  });
+}
+
 export function getMetroDirectBundleOptions(
   options: ExpoMetroOptions
 ): Partial<ExpoMetroBundleOptions> {
@@ -97,6 +112,7 @@ export function getMetroDirectBundleOptions(
     preserveEnvVars,
     asyncRoutes,
     baseUrl,
+    routerRoot,
     isExporting,
   } = withDefaults(options);
 
@@ -141,6 +157,7 @@ export function getMetroDirectBundleOptions(
       asyncRoutes,
       environment,
       baseUrl,
+      routerRoot,
     },
     customResolverOptions: {
       __proto__: null,
@@ -158,6 +175,18 @@ export function getMetroDirectBundleOptions(
   return bundleOptions;
 }
 
+export function createBundleUrlPathFromExpoConfig(
+  projectRoot: string,
+  exp: ExpoConfig,
+  options: Omit<ExpoMetroOptions, 'baseUrl' | 'routerRoot'>
+): string {
+  return createBundleUrlPath({
+    ...options,
+    baseUrl: getBaseUrlFromExpoConfig(exp),
+    routerRoot: getRouterDirectoryModuleIdWithManifest(projectRoot, exp),
+  });
+}
+
 export function createBundleUrlPath(options: ExpoMetroOptions): string {
   const {
     platform,
@@ -173,6 +202,7 @@ export function createBundleUrlPath(options: ExpoMetroOptions): string {
     preserveEnvVars,
     asyncRoutes,
     baseUrl,
+    routerRoot,
     isExporting,
   } = withDefaults(options);
 
@@ -205,6 +235,9 @@ export function createBundleUrlPath(options: ExpoMetroOptions): string {
   }
   if (baseUrl) {
     queryParams.append('transform.baseUrl', baseUrl);
+  }
+  if (routerRoot != null) {
+    queryParams.append('transform.routerRoot', routerRoot);
   }
 
   if (environment) {