+++ packages/babel-preset-expo/src/__tests__/expo-router-plugin.test.ts
@@ -27,9 +27,9 @@ function getCaller(props: Record<string, string | boolean>): babel.TransformCall
 
 beforeEach(() => {
   jest.mocked(getConfig).mockClear();
+
   process.env._EXPO_INTERNAL_TESTING = '1';
   delete process.env.EXPO_ROUTER_ABS_APP_ROOT;
-  delete process.env.EXPO_ROUTER_APP_ROOT_2;
   delete process.env.EXPO_ROUTER_IMPORT_MODE;
   delete process.env.EXPO_PROJECT_ROOT;
 });
@@ -57,22 +57,7 @@ const DEF_OPTIONS = {
   }),
 };
 it(`inlines static mode`, () => {
-  // All of this code should remain intact.
-  const sourceCode = `
-process.env.EXPO_ROUTER_IMPORT_MODE;
-export default process.env.EXPO_ROUTER_IMPORT_MODE;
-const mode = process.env.EXPO_ROUTER_IMPORT_MODE;
-  `;
-
-  expect(babel.transform(sourceCode, DEF_OPTIONS)!.code).toEqual(`
-"sync";
-export default "sync";
-const mode = "sync";`);
-
-  expect(getConfig).toHaveBeenCalledTimes(0);
-});
-
-it(`inlines async mode`, () => {
+  process.env.NODE_ENV = 'development';
   const options = {
     ...DEF_OPTIONS,
     caller: getCaller({
@@ -171,42 +156,16 @@ false;
 });
 
 it(`uses custom app entry`, () => {
-  jest
-    .mocked(getConfig)
-    .mockClear()
-    .mockReturnValueOnce({
-      pkg: {},
-      dynamicConfigObjectType: '',
-      dynamicConfigPath: '',
-      rootConfig: { expo: { slug: '...', name: '...' } },
-      staticConfigPath: '',
-      hasUnusedStaticConfig: false,
-      exp: {
-        name: '...',
-        slug: '...',
-        extra: {
-          router: { unstable_src: '/random/value' },
-        },
-      },
-    });
-
   process.env.NODE_ENV = 'development';
 
   const options = {
-    babelrc: false,
-    presets: [],
-    plugins: [expoRouterBabelPlugin],
-    sourceMaps: true,
-    filename: 'unknown',
-    configFile: false,
-    compact: false,
-    comments: true,
-    retainLines: true,
+    ...DEF_OPTIONS,
     caller: getCaller({
       name: 'metro',
       engine: 'hermes',
       projectRoot: '/foo/bar',
       platform: 'ios',
+      routerRoot: '/random/value',
     }),
   };
 
@@ -226,4 +185,40 @@ process.env.EXPO_ROUTER_APP_ROOT;`;
 "/random/value";
 // EXPO_ROUTER_APP_ROOT
 "../../../random/value";`);
+
+  expect(getConfig).toHaveBeenCalledTimes(0);
+});
+
+it(`uses custom relative app entry`, () => {
+  process.env.NODE_ENV = 'development';
+
+  const options = {
+    ...DEF_OPTIONS,
+    caller: getCaller({
+      name: 'metro',
+      engine: 'hermes',
+      projectRoot: '/foo/bar',
+      platform: 'ios',
+      routerRoot: './random/value',
+    }),
+  };
+
+  // All of this code should remain intact.
+  const sourceCode = `
+// EXPO_PROJECT_ROOT
+process.env.EXPO_PROJECT_ROOT;
+// EXPO_ROUTER_ABS_APP_ROOT
+process.env.EXPO_ROUTER_ABS_APP_ROOT;
+// EXPO_ROUTER_APP_ROOT
+process.env.EXPO_ROUTER_APP_ROOT;`;
+
+  expect(babel.transform(sourceCode, options)!.code).toEqual(`
+// EXPO_PROJECT_ROOT
+"/foo/bar";
+// EXPO_ROUTER_ABS_APP_ROOT
+"/foo/bar/random/value";
+// EXPO_ROUTER_APP_ROOT
+"../random/value";`);
+
+  expect(getConfig).toHaveBeenCalledTimes(0);
 });